+++ android/history.md
@@ -4,14 +4,15 @@
 * Start version 13.0
 * New Features:
   * Adding a download manager to execute downloads in background and cleanup the existing implementation (#2247, #2275, #2308)
+  * Show spinner (without blocking UI), if user wants to add a language/keyboard and catalog download is in progress (#2313)
   * Improve custom package installation: Show readme.htm before starting installation process (#2286)
   * Update target Android SDK version to 29 (#2279)
   * Add linting to Debug builds and resolve lint errors (#2305)
+  * Sanitize the app version to `#.#.#` for the API cloud query (#2319)
   * Check for keyboard updates during keyman startup (#2335)
   * Show available keyboard updates as android system notifications (#2335)
   * Add update indicator icon to inform user about updates and install updates in keyman app (#2335)
   
-
 ## 2019-10-30 12.0.4206 stable
 * Bug fix:
   * Disable suggestions when system keyboard entering password field (#2255)