+++ windows/src/developer/history.md
@@ -3,6 +3,12 @@
 ## 13.0 alpha
 * Start version 13.0
 
+## 2019-11-15 12.0.54 stable
+* Bug Fix: Text editor and other controls did not receive focus correctly (#2331)
+
+## 2019-10-07 12.0.50 stable
+* Release 12.0
+
 ## 2019-10-05 12.0.45 beta
 * Touch Layout Editor: Make default padding in touch layout editor match default padding in KeymanWeb. (#2170)